% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_activity_profile.R
\name{add_activity_profile}
\alias{add_activity_profile}
\title{Add an activity profile column to the dataset}
\usage{
add_activity_profile(
  data,
  activity_col,
  group_col,
  grid_length = 100,
  quantiles = c(0.05, 0.95)
)
}
\arguments{
\item{data}{dataset containing the activity counts for each group in the sample}

\item{activity_col}{the name of the column that contains the activity data}

\item{group_col}{the name of the column that contains the grouping label}
}
\value{
the same dataset with an additional columns for the activity profiles and analysis
}
\description{
Takes all of the subject's activity counts, Xt, and transforms it
into the activity profile, Ta
}
