% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_helper_columns.R
\name{add_helper_columns}
\alias{add_helper_columns}
\title{Add columns that are helpful in calculation of EL test statistic}
\usage{
add_helper_columns(data, group_col)
}
\arguments{
\item{data}{tibble containing the activity data and profiles}

\item{group_col}{string containing the name of the column with the group 
label}
}
\value{
tibble containing extra helper columns "gamma" and "n"
}
\description{
Add columns that are helpful in calculation of EL test statistic
}
