% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_U_star.R
\name{bootstrap_U_star}
\alias{bootstrap_U_star}
\title{Use bootstrap to calculate approximation of EL statistic}
\usage{
bootstrap_U_star(data, n_boot = 1000)
}
\arguments{
\item{data}{the dataset containing the activity profiles}

\item{n_boot}{how many bootstrap datasets we want to use}
}
\value{
a tibble with the results of all of the bootstrapping
}
\description{
Function uses bootstrap to calculate the uniform approximation
of the EL statistic U^2 described in Chang et. al. This function 
is used inside EL_test
}
