% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_activity_profile.R
\name{create_activity_profile}
\alias{create_activity_profile}
\title{Creates activity profile from activity counts}
\usage{
create_activity_profile(activity, activity_grid)
}
\arguments{
\item{activity_grid}{vector ofactivity counts derived from extract_activity_grid}

\item{Xt}{vector of activity counts for a subject}
}
\value{
a vector representing the activity profile for the data
}
\description{
Creates activity profile from activity counts
}
