% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_model.R
\name{ee_model}
\alias{ee_model}
\title{Estimating equation used in optimization of numerator in likelihood ratio}
\usage{
ee_model(lambda_gamma, data)
}
\arguments{
\item{lambda_gamma}{vector of parameters in equation, lambda followed by gamma}

\item{data}{dataset containing the activity profiles}
}
\value{
a vector same length as lambda_gamma that optimizes the equation
}
\description{
Estimating equation used in optimization of numerator in likelihood ratio
}
