% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EL_test.R
\name{EL_test}
\alias{EL_test}
\title{Calculates functional ANOVA test on activity profiles}
\usage{
EL_test(
  data,
  id_col,
  group_col,
  activity_col,
  grid_length = 100,
  n_boot = 1000,
  quantiles = c(0.05, 0.95),
  alpha = 0.95,
  verbose = FALSE
)
}
\arguments{
\item{data}{dataset containing the activity profiles}

\item{id_col}{the column that contains the subject ID column}

\item{group_col}{the column that contains the group ID column}

\item{activity_col}{the column that contains the activity counts column}

\item{grid_length}{number describing how long the activity grid is}

\item{n_boot}{number describing how many bootstrap datasets to make}

\item{quantiles}{two percentiles describing what percentiles of activity points to keep in grid}

\item{alpha}{number describing what quantile to check for the boostrapped test statistics}

\item{verbose}{TRUE/FALSE describing if you want to see the progresss of the functions}
}
\value{
a list contaning the results of the functional ANOVA test
}
\description{
Calculates functional ANOVA test on activity profiles
}
\examples{
params_by_group = list(
  list(n_subj = 70, n_points = 1000, 
      scaling = 300, thetas = c(-0.8, 2.0, 2.31)),
 list(n_subj = 100, n_points = 1000, 
      scaling = 300, thetas = c(-0.402, 1.0, 1.65)),
 list(n_subj = 130, n_points = 1000, 
      scaling = 300, thetas = c(-0.201, 0.5, 1.18))
 )

data = prepare_sim_data(params_by_group)
el = EL_test(data, id_col = subject_id, group_col = group, activity_col = Xt)

}
