% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_activity}
\alias{example_activity}
\title{Simulated dataset}
\description{
These data are included in the package as an example. 
They were generated using a package that is no longer available, although
code for generating the data can be provided.
}
\keyword{data}
