% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_activity_grid.R
\name{extract_activity_grid}
\alias{extract_activity_grid}
\title{Create a grid of activity values to standardize activity profiles}
\usage{
extract_activity_grid(data, grid_length = 100, quantiles = c(0.05, 0.95))
}
\arguments{
\item{data}{dataset containing groups and activity}

\item{grid_length}{number describing how long do we want the activity grid to be}

\item{quantiles}{two percentiles describing what percentiles of activity points to keep in grid}
}
\value{
a vector of activity values
}
\description{
Takes a dataset containing activity counts for multiple groups
and creates a standard grid/vector of activity counts that can 
be applied across all of the groups in the dataset. Activity counts
assumed to be nested. See output of prepare_sim_data
}
\examples{
activity_grid = extract_activity_grid(data)
}
