% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate activity data}
\usage{
generate_data(n_subj, group_id, n_points, scaling, thetas)
}
\arguments{
\item{n_subj}{number describing how many subjects are in this particular group}

\item{group_id}{number describing the index of the group to differentiate it from others}

\item{n_points}{how much activity count data we want to produce for each person}

\item{scaling}{how much the activity data should be scaled}

\item{thetas}{numbers for parameters in creating Ornstein-Uhlenbeck process}
}
\value{
data containing the activity data
}
\description{
Simulates data for a group of subjects using an Ornstein-Uhlenbeck process
with user defined parameters for the process
}
\examples{
# Generate data for a single group of subjects
generate_data(n_subj = 70, group_id = 1, n_points = 1000, 
              scaling = 300, thetas = c(-0.8, 2.0, 2.31))
}
