% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_grid.R
\name{initialize_grid}
\alias{initialize_grid}
\title{Creates grid of mu and lambda values to optimize estimation equation}
\usage{
initialize_grid(
  data,
  mu_grid_length = 20,
  lambda_grid_length = 20,
  mu_tolerance = 20,
  lambda_tolerance = 20
)
}
\arguments{
\item{data}{dataset containing the activity profiles}

\item{mu_grid_length}{number describing how long the mu grid should be}

\item{lambda_grid_length}{number describing how long the lambda grid should be}

\item{mu_tolerance}{number describing how fine the different mus should be in the grid}

\item{lambda_tolerance}{number describing how fine the different mus should be in the grid}
}
\value{
a tibble with all of the valid mu-lambda combinations from the data
}
\description{
Creates grid of mu and lambda values to optimize estimation equation
}
