% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg2logRa.R
\name{neg2logRa}
\alias{neg2logRa}
\title{Calculates -2logR(a) for a given activity index for data}
\usage{
neg2logRa(data, a, verbose = FALSE)
}
\arguments{
\item{data}{dataset containing activity profiles}

\item{a}{the activity index that we want to calculate -2logR(a) at}

\item{verbose}{TRUE/FALSE describing if you want to see progress of function}
}
\value{

}
\description{
Calculates -2logR(a) for a given activity index for data
}
