% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sim_data.R
\name{prepare_sim_data}
\alias{prepare_sim_data}
\title{Simulate activity data for a groups of subjects}
\usage{
prepare_sim_data(group_params, quantiles = c(0.05, 0.95))
}
\arguments{
\item{group_params}{A list of parameters for each group in the simulated sample}

\item{quantiles}{A vector of two percentages describing how which quantiles of activity to keep}
}
\value{
A tibble of the activity for the sample
}
\description{
Given parameters for each group, simulates activity data for each group.
}
\examples{
params_by_group = list(
  list(n_subj = 70, n_points = 1000, 
      scaling = 300, thetas = c(-0.8, 2.0, 2.31)),
 list(n_subj = 100, n_points = 1000, 
      scaling = 300, thetas = c(-0.402, 1.0, 1.65)),
 list(n_subj = 130, n_points = 1000, 
      scaling = 300, thetas = c(-0.201, 0.5, 1.18))
 )

data = prepare_sim_data(params_by_group)

}
