% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_activity_profile.R
\name{scale_activity_profile}
\alias{scale_activity_profile}
\title{Scales activity profiles so that max range is 1}
\usage{
scale_activity_profile(data, a)
}
\arguments{
\item{data}{dataset containing the activity profiles}

\item{a}{activity index that describes where in the activity profiles to filter}
}
\value{
data with another column containing the scaled activity profiles
}
\description{
Scales activity profiles so that max range is 1
}
