% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_activity_profile.R
\name{add_activity_profile}
\alias{add_activity_profile}
\title{Add an activity profile column to the dataset}
\usage{
add_activity_profile(data, activity_grid)
}
\arguments{
\item{data}{dataset containing the activity counts for each group in the sample}

\item{activity_grid}{A vector derived from extract_activity_grid}
}
\value{
the same dataset with an additional columns for the activity profiles and analysis
}
\description{
Takes all of the subject's activity counts, Xt, and transforms it
into the activity profile, Ta
}
\examples{

activity_grid = extract_activity_grid(data)
processed_data = add_activity_profile(data, activity_grid)

}
