% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_activity_grid.R
\name{extract_activity_grid}
\alias{extract_activity_grid}
\title{Given a matrix of activity data, output a standardized equidistant vector of  
activity values that is compatible with all values in the data}
\usage{
extract_activity_grid(
  data,
  activity_col,
  group_col,
  grid_length = 100,
  quantiles = c(0.05, 0.95)
)
}
\arguments{
\item{data}{dataset containing groups and activity}

\item{activity_col}{the name of the column that contains the activity data}

\item{group_col}{the name of the column that contains the grouping label}

\item{grid_length}{number describing how long the activity grid should be}

\item{quantiles}{two percentiles describing what percentiles of activity 
points to keep in grid}
}
\value{
a vector of equidistant activity values from the maximum to the 
minimum in the data
}
\description{
Takes a dataset containing activity counts for multiple groups
and creates a standard grid/vector of activity counts that can 
be applied across all of the groups in the dataset. Activity counts
assumed to be nested. See output of prepare_sim_data
}
